fix(book): 가격 타입 변경(int -> BigDecimal)

+++ src/main/java/com/nhnacademy/illuwa/cart/dto/CartResponse.java
@@ -20,7 +20,7 @@ public class CartResponse {
         this.cartId = cartId;
         this.bookCarts = bookCarts;
         this.totalPrice = bookCarts.stream()
-                .map(item -> BigDecimal.valueOf(item.getSalePrice()).multiply(BigDecimal.valueOf(item.getAmount())))
+                .map(item -> item.getSalePrice().multiply(BigDecimal.valueOf(item.getAmount())))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 }
